ID's are now longs. Next id is saved to a file so that it can be reloaded and main will now reload files correctly from the individual folders. Most of this main code should be moved and wrapped in tests now that it's getting fairly complicated to load data.

+++ src/main/java/com/andrewauclair/todo/Commands.java
@@ -78,7 +78,7 @@ public class Commands {
 		catch (RuntimeException ignored) {
 		}
 		
-		final int activeTaskID = activeTask == null ? -1 : activeTask.id;
+		final long activeTaskID = activeTask == null ? -1 : activeTask.id;
 		
 		tasksList.stream()
 				.limit(20)
